pyproject.toml
[project]
name = "organize-desktop"
version = "1.0.0"
description = "A modern desktop application for the organize file management automation tool"
authors = [{ name = "Organize Desktop Team", email = "contact@organize-desktop.app" }]
license = { text = "MIT" }
readme = "README.md"
requires-python = ">=3.9"
keywords = [
    "file",
    "management",
    "automation",
    "desktop",
    "gui",
    "organize",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: X11 Applications :: Qt",
    "Intended Audience :: End Users/Desktop",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Topic :: Desktop Environment :: File Managers",
    "Topic :: Utilities",
]
dependencies = [
    "organize-tool>=3.3.0,<4.0.0",
    "PyQt6>=6.6.0,<7.0.0",
    "PyQt6-Charts>=6.6.0,<7.0.0",
    "pyyaml>=6.0,<7.0",
    "platformdirs>=4.0.0,<5.0.0",
    "watchdog>=3.0.0,<5.0.0",
    "qtawesome>=1.3.0,<2.0.0",
    "qdarkstyle>=3.2.0,<4.0.0",
    "pyqtgraph>=0.13.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0,<9.0.0",
    "pytest-qt>=4.2.0,<5.0.0",
    "ruff>=0.3.0,<1.0.0",
    "mypy>=1.0.0,<2.0.0",
]

[project.scripts]
organize-desktop = "app.main:main"

[project.gui-scripts]
organize-desktop-gui = "app.main:main"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
